requirements/pdf.txt
PyPDF2>=3.0.0
fpdf2>=2.7.0
